Готово

+++ TSSA_LABS/src/net/zvegincev/LB1_availability_step1.java
@@ -6,7 +6,7 @@ public class LB1_availability_step1 {
 		
 	}
 	
-	public boolean availability(float X, float Y, float conditionX1, float conditionX2, float conditionY1, float conditionY2) {
+	public boolean availability(double X, double Y, double conditionX1, double conditionX2, double conditionY1, double conditionY2) {
         if(X>=conditionX1 && X<=conditionX2){
         	if(Y>=conditionY1 && Y<=conditionY2){
         		return true;
